Added API encryption to API endpoints

+++ CallejoIncChildCareAPI/Controllers/ExpenseController.cs
@@ -4,6 +4,8 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Common.Services.Expenses;
 using Microsoft.Extensions.Configuration.UserSecrets;
+using Common.Services.Login;
+using CallejoIncChildCareAPI.Authorize;
 
 namespace CallejoIncChildcareAPI.Controllers
 {
@@ -13,23 +15,34 @@ namespace CallejoIncChildcareAPI.Controllers
     public class ExpenseController : ControllerBase
     {
         private readonly IExpenseService _expenseService;
+        private ILoginService _loginService;
 
-        public ExpenseController(IExpenseService expenseService)
+        public ExpenseController(IExpenseService expenseService, ILoginService loginService)
         {
             _expenseService = expenseService;
+            _loginService = loginService;
         }
 
-
+        [AuthorizeAttribute()]
         [HttpGet("children/count")]
         public async Task<IActionResult> GetChildrenCount()
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
             int count = await _expenseService.GetChildrenCountAsync();
             return Ok(count); // count is a simple int, easily serializable
         }
 
+        [AuthorizeAttribute()]
         [HttpGet("expenses/total")]
         public async Task<IActionResult> GetTotalExpenses()
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
             var total = await _expenseService.GetTotalExpensesAsync();
             return Ok(total);
         }
@@ -37,6 +50,7 @@ namespace CallejoIncChildcareAPI.Controllers
 
 
         // POST: api/Expenses/Upload
+        [AuthorizeAttribute()]
         [HttpPost("Upload")]
         public async Task<ActionResult<ExpenseDTO>> UploadExpense(
             IFormFile file,
@@ -45,6 +59,10 @@ namespace CallejoIncChildcareAPI.Controllers
             [FromForm] decimal amount,
             [FromForm] string? note)
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
             if (file == null || file.Length == 0)
             {
                 return BadRequest("Receipt file is required.");
@@ -84,9 +102,14 @@ namespace CallejoIncChildcareAPI.Controllers
             return CreatedAtAction(nameof(UploadExpense), new {id =  createdExpense.Id}, createdExpense);
         }
 
+        [AuthorizeAttribute()]
         [HttpDelete("Delete")]
         public async Task<IActionResult> DeleteExpense(int id)
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
             var deleted = await _expenseService.DeleteExpenseAsync(id);
 
             if (!deleted)
@@ -97,6 +120,7 @@ namespace CallejoIncChildcareAPI.Controllers
             return NoContent(); // HTTP 204 -- deletion successful
         }
 
+        [AuthorizeAttribute()]
         [HttpPut("Edit")]
         public async Task<IActionResult> EditExpense(
             int id,
@@ -106,6 +130,10 @@ namespace CallejoIncChildcareAPI.Controllers
             [FromForm] decimal amount,
             [FromForm] string? note)
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
             if (!DateOnly.TryParse(date, out DateOnly parsedDate))
             {
                 return BadRequest("Invalid date format. Use MM/DD/YYYY");
